# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loop API",
    "version": "0.1.0",
    "description": "The Loop Autonomous Improvement Engine API. Collects signals, organizes work into issues, and dispatches AI agents with hydrated prompts.",
    "contact": {
      "name": "Loop by Dork Labs",
      "url": "https://www.looped.me"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5667",
      "description": "Local development"
    },
    {
      "url": "https://api.looped.me",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token in the Authorization header"
      }
    },
    "schemas": {
...
      "Label": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "name": {
            "type": "string",
            "example": "frontend"
          },
          "color": {
            "type": "string",
            "example": "#ff6b6b"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": ["id", "name", "color", "createdAt", "deletedAt"]
      },
...
      "Goal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "title": {
            "type": "string",
            "example": "Reduce login failure rate to < 5%"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Track auth failure rates from PostHog."
          },
          "metric": {
            "type": ["string", "null"],
            "example": "login_failure_rate"
          },
          "targetValue": {
            "type": ["number", "null"],
            "example": 0.05
          },
          "currentValue": {
            "type": ["number", "null"],
            "example": 0.18
          },
          "unit": {
            "type": ["string", "null"],
            "example": "ratio"
          },
          "status": {
            "type": "string",
            "enum": ["active", "achieved", "abandoned"],
            "example": "active"
          },
          "projectId": {
            "type": ["string", "null"],
            "example": "cuid2_abc123"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "metric",
          "targetValue",
          "currentValue",
          "unit",
          "status",
          "projectId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "name": {
            "type": "string",
            "example": "Auth Reliability"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Improve authentication reliability and reduce failures."
          },
          "status": {
            "type": "string",
            "enum": ["backlog", "planned", "active", "paused", "completed", "canceled"],
            "example": "active"
          },
          "health": {
            "type": "string",
            "enum": ["on_track", "at_risk", "off_track"],
            "example": "on_track"
          },
          "goalId": {
            "type": ["string", "null"],
            "example": "cuid2_abc123"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "status",
          "health",
          "goalId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
...
    },
...
  },
  "paths": {
...
    "/api/labels": {
      "get": {
        "tags": ["Labels"],
        "summary": "List labels",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of labels",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Label"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Labels"],
        "summary": "Create a label",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 100,
                    "example": "frontend"
                  },
                  "color": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50,
                    "example": "#ff6b6b"
                  }
                },
                "required": ["name", "color"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Label created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Label"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "409": {
            "description": "Label name already exists"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
    "/api/labels/{id}": {
      "delete": {
        "tags": ["Labels"],
        "summary": "Soft-delete a label",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Label deleted"
          },
          "404": {
            "description": "Label not found"
          }
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List projects",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of projects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Project"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "Auth Reliability"
                  },
                  "description": {
                    "type": "string",
                    "example": "Improve authentication reliability."
                  },
                  "status": {
                    "type": "string",
                    "enum": ["backlog", "planned", "active", "paused", "completed", "canceled"],
                    "default": "backlog",
                    "example": "active"
                  },
                  "health": {
                    "type": "string",
                    "enum": ["on_track", "at_risk", "off_track"],
                    "default": "on_track",
                    "example": "on_track"
                  },
                  "goalId": {
                    "type": "string"
                  }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Project"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "422": {
            "description": "Validation error or goal not found"
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project by ID",
        "description": "Returns the project with linked goal and issue counts by status.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Project with goal and issue counts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Project"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "goal": {
                              "allOf": [
                                {
                                  "$ref": "#/components/schemas/Goal"
                                },
                                {
                                  "type": ["object", "null"]
                                }
                              ]
                            },
                            "issueCounts": {
                              "type": "object",
                              "additionalProperties": {
                                "type": "number"
                              },
                              "example": {
                                "todo": 3,
                                "in_progress": 1,
                                "done": 10
                              }
                            }
                          },
                          "required": ["goal", "issueCounts"]
                        }
                      ]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          }
        }
      },
      "patch": {
        "tags": ["Projects"],
        "summary": "Update a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "description": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["backlog", "planned", "active", "paused", "completed", "canceled"]
                  },
                  "health": {
                    "type": "string",
                    "enum": ["on_track", "at_risk", "off_track"]
                  },
                  "goalId": {
                    "type": ["string", "null"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated project",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Project"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "404": {
            "description": "Project not found"
          },
          "422": {
            "description": "Validation error or goal not found"
          }
        }
      },
      "delete": {
        "tags": ["Projects"],
        "summary": "Soft-delete a project",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted"
          },
          "404": {
            "description": "Project not found"
          }
        }
      }
    },
...
  },
...
}